main sort values

+++ model/src/com/constellio/model/services/records/cache/dataStore/RecordsCachesDataStore.java
@@ -3,6 +3,7 @@ package com.constellio.model.services.records.cache.dataStore;
 import com.constellio.data.dao.dto.records.RecordDTO;
 import com.constellio.data.utils.CacheStat;
 import com.constellio.data.utils.LazyMergingIterator;
+import com.constellio.data.utils.dev.Toggle;
 import com.constellio.model.conf.FoldersLocator;
 import com.constellio.model.services.factories.ModelLayerFactory;
 import com.constellio.model.services.records.RecordId;
@@ -51,41 +52,43 @@ public class RecordsCachesDataStore {
 		this.stringIdsDataStore = new StringIdsRecordsCachesDataStore(modelLayerFactory);
 
 
-		List<RecordId> recordIds = null;
-		File idsList = new File(new FoldersLocator().getWorkFolder(), "integer-ids.txt");
-		if (FoldersLocator.usingAppWrapper()) {
-			if (idsList.exists() && new DateTime(idsList.lastModified()).isAfter(new DateTime().minusDays(1))) {
-				try {
-					recordIds = FileUtils.readLines(idsList, "UTF-8").stream().map((line) -> RecordId.toId(line))
-							.filter((id) -> id.isInteger()).collect(Collectors.toList());
-				} catch (IOException e) {
-					e.printStackTrace();
-					recordIds = null;
-				}
-			}
-		}
-		if (recordIds == null) {
-			LOGGER.info("Loading ids from solr... could take up to 30 minutes, please wait...");
-			Iterator<RecordId> recordIdIterator = modelLayerFactory.newSearchServices().recordsIdIteratorExceptEvents();
-			recordIds = IteratorUtils.toList(recordIdIterator);
-			List<String> lines = recordIds.stream().filter((id) -> id.isInteger()).map(RecordId::stringValue).collect(Collectors.toList());
-
+		if (Toggle.STRUCTURE_CACHE_BASED_ON_EXISTING_IDS.isEnabled()) {
+			List<RecordId> recordIds = null;
+			File idsList = new File(new FoldersLocator().getWorkFolder(), "integer-ids.txt");
 			if (FoldersLocator.usingAppWrapper()) {
-				try {
-
-					FileUtils.writeLines(idsList, lines);
-					if (!lines.isEmpty()) {
-						LOGGER.info("Last line is : " + lines.get(lines.size() - 1));
+				if (idsList.exists() && new DateTime(idsList.lastModified()).isAfter(new DateTime().minusDays(1))) {
+					try {
+						recordIds = FileUtils.readLines(idsList, "UTF-8").stream().map((line) -> RecordId.toId(line))
+								.filter((id) -> id.isInteger()).collect(Collectors.toList());
+					} catch (IOException e) {
+						e.printStackTrace();
+						recordIds = null;
 					}
-				} catch (IOException e) {
-					throw new RuntimeException(e);
 				}
 			}
-		}
+			if (recordIds == null) {
+				LOGGER.info("Loading ids from solr... could take up to 30 minutes, please wait...");
+				Iterator<RecordId> recordIdIterator = modelLayerFactory.newSearchServices().recordsIdIteratorExceptEvents();
+				recordIds = IteratorUtils.toList(recordIdIterator);
+				List<String> lines = recordIds.stream().filter((id) -> id.isInteger()).map(RecordId::stringValue).collect(Collectors.toList());
 
-		LOGGER.info("Structuring cache based on ids...       - Current memory : " + humanReadableByteCount(OffHeapMemoryAllocator.getAllocatedMemory(), true));
-		intIdsDataStore.structureCacheUsingExistingIds(recordIds.iterator());
-		LOGGER.info("Structuring cache based on ids finished - Current memory : " + humanReadableByteCount(OffHeapMemoryAllocator.getAllocatedMemory(), true));
+				if (FoldersLocator.usingAppWrapper()) {
+					try {
+
+						FileUtils.writeLines(idsList, lines);
+						if (!lines.isEmpty()) {
+							LOGGER.info("Last line is : " + lines.get(lines.size() - 1));
+						}
+					} catch (IOException e) {
+						throw new RuntimeException(e);
+					}
+				}
+			}
+
+			LOGGER.info("Structuring cache based on ids...       - Current memory : " + humanReadableByteCount(OffHeapMemoryAllocator.getAllocatedMemory(), true));
+			intIdsDataStore.structureCacheUsingExistingIds(recordIds.iterator());
+			LOGGER.info("Structuring cache based on ids finished - Current memory : " + humanReadableByteCount(OffHeapMemoryAllocator.getAllocatedMemory(), true));
+		}
 	}
 
 	public void insertWithoutReservingSpaceForPreviousIds(RecordDTO dto) {
@@ -277,5 +280,20 @@ public class RecordsCachesDataStore {
 	public void structureCacheUsingExistingIds(Iterator<RecordId> existingIdsIterator) {
 		intIdsDataStore.structureCacheUsingExistingIds(existingIdsIterator);
 
+	}
+
+	public void setRecordsMainSortValue(List<RecordId> existingIds) {
+		intIdsDataStore.setRecordsMainSortValue(existingIds);
+		stringIdsDataStore.setRecordsMainSortValue(existingIds);
+
+	}
+
+	public int getMainSortValue(RecordId recordId) {
+		if (recordId.isInteger()) {
+			return intIdsDataStore.getMainSortValue(recordId);
+
+		} else {
+			return -1;
+		}
 	}
 }
